BUGFIX: Casting pointers broken for 64bit platforms. Used as session ID.
git-svn-id: svn://svn.code.sf.net/p/nel/code/trunk@235 bf685bcc-7839-4fcd-9254-36823579bede

+++ nel/samples/net/udp/bench_service.cpp
@@ -166,7 +166,7 @@ string getDate()
 
 void cbInit (CMessage &msgin, TSockId from, CCallbackNetBase &netbase)
 {
-	uint32 session = (uint32) from;
+	uint64 session = (uint64)(uintptr_t) from;
 
 	string connectionName;
 	msgin.serial (connectionName);
